+++ .github/workflows/no-default-features.yml
@@ -15,20 +15,13 @@ on:
 jobs:
   check:
     runs-on: windows-2022
-
-    strategy:
-      matrix:
-        include:
-          - version: nightly
-            target: x86_64-pc-windows-msvc
-
     steps:
       - name: Checkout
         uses: actions/checkout@v4
       - name: Update toolchain
-        run: rustup update --no-self-update ${{ matrix.version }} && rustup default ${{ matrix.version }}-${{ matrix.target }}
+        run: rustup update --no-self-update nightly && rustup default nightly-x86_64-pc-windows-msvc
       - name: Add toolchain target
-        run: rustup target add ${{ matrix.target }}
+        run: rustup target add x86_64-pc-windows-msvc
       - name: Fix environment
         uses: ./.github/actions/fix-environment
       - name: Check cppwinrt
