+++ crates/tools/yml/src/clippy.rs
@@ -1,0 +1,54 @@
+use super::*;
+
+pub fn yml() {
+    let mut yml = r"name: clippy
+
+on:
+  pull_request:
+    paths-ignore:
+      - '.github/ISSUE_TEMPLATE/**'
+      - 'web/**'
+  push:
+    paths-ignore:
+      - '.github/ISSUE_TEMPLATE/**'
+      - 'web/**'
+    branches:
+      - master
+
+jobs:
+  check:
+    runs-on: windows-2022
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+      - name: Update toolchain
+        run: rustup update --no-self-update nightly && rustup default nightly-x86_64-pc-windows-msvc
+      - name: Add toolchain target
+        run: rustup target add x86_64-pc-windows-msvc
+      - name: Install clippy
+        run: rustup component add clippy
+      - name: Fix environment
+        uses: ./.github/actions/fix-environment"
+        .to_string();
+
+    // This unrolling is required since "cargo clippy --all" consumes too much memory for the GitHub hosted runners.
+
+    for package in helpers::crates("crates") {
+        let name = &package.name;
+
+        // Tests are skipped since there's a lot of noise from some tests that are not useful/relevant to correct.
+        if name.starts_with("test_") {
+            continue;
+        }
+
+        write!(
+            &mut yml,
+            r"
+      - name: Check {name}
+        run:  cargo clippy -p {name}"
+        )
+        .unwrap();
+    }
+
+    std::fs::write(".github/workflows/clippy.yml", yml.as_bytes()).unwrap();
+}
